backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.0
h11==0.16.0
idna==3.10
packaging==25.0
pipx==1.7.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
userpath==1.9.2
uvicorn==0.35.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0